Vector2D float to double

+++ src/utility/Vector2D.hpp
@@ -5,15 +5,15 @@
 
 class Vector2D {
     public:
-        Vector2D(float x, float y): m_x(x), m_y(y) {}
+        Vector2D(double x, double y): m_x(x), m_y(y) {}
 
-        float getX() const { return m_x; }
-        float getY() const { return m_y; }
+        double getX() const { return m_x; }
+        double getY() const { return m_y; }
 
-        void setX(float x) { m_x = x; }
-        void setY(float y) { m_y = y; }
+        void setX(double x) { m_x = x; }
+        void setY(double y) { m_y = y; }
 
-        float length() const { return sqrt(m_x * m_x + m_y * m_y); }
+        double length() const { return sqrt(m_x * m_x + m_y * m_y); }
 
         Vector2D operator+ (const Vector2D& v2) const {
             return Vector2D(m_x + v2.m_x, m_y + v2.m_y);
@@ -25,11 +25,11 @@ class Vector2D {
             return v1;
         }
 
-        Vector2D operator* (float scalar) {
+        Vector2D operator* (double scalar) {
             return Vector2D(m_x * scalar, m_y * scalar);
         }
 
-        Vector2D& operator*= (float scalar) {
+        Vector2D& operator*= (double scalar) {
             m_x *= scalar;
             m_y *= scalar;
             return *this;
@@ -45,26 +45,26 @@ class Vector2D {
             return v1;
         }
 
-        Vector2D operator/(float scalar) {
+        Vector2D operator/(double scalar) {
             return Vector2D(m_x / scalar, m_y / scalar);
         }
 
-        Vector2D& operator/= (float scalar) {
+        Vector2D& operator/= (double scalar) {
             m_x /= scalar;
             m_y /= scalar;
             return *this;
         }
 
         void normalize() {
-            float l = length();
+            double l = length();
             if (l > 0) {
                 (*this) *= 1/l;
             }
         }
 
     private:
-        float m_x;
-        float m_y;
+        double m_x;
+        double m_y;
 };
 
 #endif
